Change grid's size
I add the button in the html and every time the user click the button
the program ask the user to type a number. After confirm if the limit of
the grid's size is ok it changes the number's limit variable and calls a
function to create the grid again with the new limit.
I use a while to delete every first child of container until is empty.