Add progressive darkening effect
It has the same logic as rainbow and normal draws. I convert the string
to number and viceversa to get a change in the opacity, every time the
mouse over it, it changes 0.1 up to 1.
I also remove the background color and add the reset grid every time the
user click on the buttons.